requirements.txt
flask
flask_restplus
pyserial
